+++ grafika Hw2/src/objects/Token.h
@@ -16,10 +16,10 @@ private:
 	GLfloat *color;
 	int created_time_move;
 public:
-	Token(float,float,float,float,int);
+	Token(double,double,double,double,int);
 	virtual ~Token();
 	void view();
-	void setPosition(float,float,float);
+	void setPosition(double,double,double);
 	void moveForward(); /*Moves to x, or z*/
 	void moveBackWard(); /*ROtates/ ,Update dir z and call moveFoward*/
 	void moveLeft();	/*Rotate update dir x and call movefoward*/
